* kiklib/src/Makefile.in: MAJOR.MINOR 14.2 => 14.3 * kik_unistd.h: #include <stdlib.h> is added for putenv. * x_font_cache.c, x_font_config.c:   font_hash => kik_map_hash_int * kik_map.[ch]:   - kik_map_hash_int_fast() is added.   - DEFAULT_MAP_SIZE is changed from 128 to 16.   - (map)->pairs_array is allocated in kik_map_get_pairs_array() instead     of kik_map_new_with_size().   - The bug which erases data unexpectedly in shrinking the map size     is fixed.

+++ xwindow/fb/x_imagelib.c
@@ -110,7 +110,7 @@ modify_pixmap(
 	u_int  y ;
 	u_int  x ;
 	u_char  r , g , b , a ;
-	u_int32_t  pixel ;
+	u_long  pixel ;
 
 	if( ! x_picture_modifier_is_normal( pic_mod) &&
 	    ( value_table = alloca( 256)))
